+++ metadata-ingestion/docs/sources/dummy-data/dummy-data_recipe.yml
@@ -1,0 +1,5 @@
+source:
+  type: dummy-data
+  config:
+    # This is set
+    path: ""
