--- metadata-ingestion/docs/sources/sample-data/sample-data_recipe.yml
@@ -1,5 +1,0 @@
-source:
-  type: sample-data
-  config:
-    # This is ignored so doesn't matter.
-    path: ""
