+++ metadata-ingestion/src/datahub/cli/docker_cli.py
@@ -29,6 +29,11 @@ from datahub.cli.docker_check import (
 from datahub.ingestion.run.pipeline import Pipeline
 from datahub.telemetry import telemetry
 from datahub.upgrade import upgrade
+from datahub.utilities.sample_data import (
+    BOOTSTRAP_MCES_FILE,
+    DOCKER_COMPOSE_BASE,
+    download_sample_data,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -48,13 +53,6 @@ ELASTIC_CONSUMERS_QUICKSTART_COMPOSE_FILE = (
     "docker/quickstart/docker-compose.consumers-without-neo4j.quickstart.yml"
 )
 
-BOOTSTRAP_MCES_FILE = "metadata-ingestion/examples/mce_files/bootstrap_mce.json"
-
-DOCKER_COMPOSE_BASE = os.getenv(
-    "DOCKER_COMPOSE_BASE",
-    "https://raw.githubusercontent.com/datahub-project/datahub/master",
-)
-
 NEO4J_AND_ELASTIC_QUICKSTART_COMPOSE_URL = (
     f"{DOCKER_COMPOSE_BASE}/{NEO4J_AND_ELASTIC_QUICKSTART_COMPOSE_FILE}"
 )
@@ -62,18 +60,6 @@ ELASTIC_QUICKSTART_COMPOSE_URL = (
     f"{DOCKER_COMPOSE_BASE}/{ELASTIC_QUICKSTART_COMPOSE_FILE}"
 )
 M1_QUICKSTART_COMPOSE_URL = f"{DOCKER_COMPOSE_BASE}/{M1_QUICKSTART_COMPOSE_FILE}"
-BOOTSTRAP_MCES_URL = f"{DOCKER_COMPOSE_BASE}/{BOOTSTRAP_MCES_FILE}"
-
-
-def download_sample_data() -> str:
-    with tempfile.NamedTemporaryFile(suffix=".json", delete=False) as tmp_file:
-        path = str(pathlib.Path(tmp_file.name))
-
-        # Download the bootstrap MCE file from GitHub.
-        mce_json_download_response = requests.get(BOOTSTRAP_MCES_URL)
-        mce_json_download_response.raise_for_status()
-        tmp_file.write(mce_json_download_response.content)
-    return path
 
 
 class Architectures(Enum):
